Django_base/requirements.txt
Django==2.2.28
python-dotenv==0.20.0
djangorestframework==3.13.1
python-nmap==0.7.1
drf-yasg2==1.19.4
PyYAML==6.0